fix appt id sent to server to reflect correct datatype and mongodb id for status change
